update use cases
