+++ mobile/lib/features/movie/domain/use_cases/get_movie_detail.dart
@@ -1,0 +1,23 @@
+import 'package:dartz/dartz.dart';
+
+import '../../../../core/error/failure.dart';
+import '../entities/movie_entity.dart';
+import '../repositories/movie_repository.dart';
+import 'use_case.dart';
+
+class GetMovieDetailUseCase extends UseCase<MovieEntity, GetMovieParams> {
+  final MovieRepository repository;
+
+  GetMovieDetailUseCase({required this.repository});
+
+  @override
+  Future<Either<Failure, MovieEntity>> call(GetMovieParams params) async {
+    return await repository.getMovie(params.movieId);
+  }
+}
+
+class GetMovieParams {
+  final String movieId;
+
+  GetMovieParams({required this.movieId});
+}
